Migration from V1.0 dialogflow API to V2 API (please reinstall if updating from a previous version)
Be careful, a lot of changes were made to this plugin in order to adapt it for V2 API. If you're trying to update from a previous version of this plugin, please make a fresh install of it in order to get all the required dependencies. Please refer to https://github.com/Potjoe-97/jarvis-dialogflow/ and consult instructions to install it.